@@ -9,8 +9,8 @@
         }
     ],
     "require": {
-        "php": ">=5.6.4",
-        "illuminate/log": "~5.4"
+        "php": ">=7.1.0",
+        "illuminate/log": "~5.6"
     },
     "autoload": {
         "psr-4": {
@@ -19,5 +19,12 @@
         "files": [
             "src/helpers.php"
         ]
+    },
+    "extra": {
+        "laravel": {
+            "providers": [
+                "Tochka\\JsonRpc\\JsonRpcServiceProvider"
+            ]
+        }
     }
 }
